app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.30.1
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
aiofiles==23.2.1
